backend/requirements.txt
# annotated-types==0.6.0
# anyio==4.3.0
# click==8.1.7
# colorama==0.4.6
# exceptiongroup==1.2.0
# fastapi==0.110.0
# h11==0.14.0
# httptools==0.6.1
# idna==3.6
# pydantic==2.6.3
# pydantic_core==2.16.3
# pyspark==3.3.1
# python-dotenv==1.0.1
# PyYAML==6.0.1
# sniffio==1.3.1
# starlette==0.36.3
# typing_extensions==4.10.0
# uvicorn==0.28.0
# watchfiles==0.21.0
# websockets==12.0
# langchain
# sentence-transformers
# chromadb
# torch==2.1.0
# accelerate==0.22.0
# bitsandbytes
langchain
openai
chromadb
tiktoken
PyPDF2
pypdf
python-docx
pandas
python-dotenv
fastapi
uvicorn
python-multipart
chardet